Crate steel_registry

Re-exports

pub use enchantment::effect as enchantment_effect;

pub use mob_effect::instance as mob_effect_instance;

pub use registry::holder;

pub use registry::holder_set;

pub use registry::reference as registry_reference;

pub use structure::processor as structure_processor;

pub use structure::set as structure_set;

pub use structure::template_pool;

pub use consume_effect::ConsumeEffectData;

pub use consume_effect::ConsumeEffectType;

pub use consume_effect::ConsumeEffectTypeRef;

pub use dye_color::DyeColor;

pub use entity_variant::AxolotlVariant;

pub use entity_variant::FoxVariant;

pub use entity_variant::HorseVariant;

pub use entity_variant::LlamaVariant;

pub use entity_variant::MooshroomVariant;

pub use entity_variant::ParrotVariant;

pub use entity_variant::RabbitVariant;

pub use entity_variant::SalmonVariant;

pub use entity_variant::TropicalFishBase;

pub use entity_variant::TropicalFishPattern;

pub use item_stack_template::ItemStackTemplate;

pub use mob_effect_instance::MobEffectInstance;

pub use mob_effect_instance::MobEffectInstanceDetails;

pub use potion::Potion;

pub use potion::PotionEffect;

pub use potion::PotionRef;

pub use resolvable_profile::PartialProfile;

pub use resolvable_profile::PlayerModelType;

pub use resolvable_profile::PlayerSkinPatch;

pub use resolvable_profile::ProfileProperty;

pub use resolvable_profile::ResolvableProfile;

pub use resolvable_profile::ResolvableProfileContents;

pub use resolvable_profile::StoredGameProfile;

pub use registry::*;

Modules

attribute

banner_pattern

Banner pattern registry values.

biome

block_entity_type

blocks

carver

Configured carver registry.

cat_sound_variant

cat_variant

chat_type

chicken_sound_variant

chicken_variant

consume_effect

Registry-dispatched effects applied after consuming or saving an item from death.

cow_sound_variant

cow_variant

damage_type

data_component_predicate

Registry-dispatched partial data-component predicates.

data_components

Data components system for items and entities.

dialog

dimension_type

dye_color

Vanilla dye colors shared by item components, entities, and blocks.

enchantment

entity_data

Entity data synchronization system.

entity_type

entity_variant

Closed Vanilla entity variants shared by item components and entities.

equipment

Shared equipment slot definitions.

feature

Configured and placed feature registries.

fluid

Fluid registry for Minecraft fluids.

frog_variant

game_events

game_rules

instrument

item_predicate

Vanilla item, block, state, and NBT predicate codec values.

item_stack

Item stack implementation.

item_stack_template

Non-empty item stack templates used by recursive Vanilla codecs.

items

jukebox_song

level_events

Level event constants matching vanilla Minecraft’s LevelEvent.java.

loot_table

macros 🔒

map_decoration_type

Map decoration type registry values extracted from Vanilla.

menu_type

mob_effect

packets

painting_variant

particle_type

pig_sound_variant

pig_variant

poi

Point of Interest (POI) type registry.

position_source

potion

Potion registry values extracted from Vanilla.

recipe

Recipe system for crafting and other recipe types.

registry

Registry ownership, lookup, tagging, and reference codecs.

resolvable_profile

Vanilla resolvable player profiles used by item components and entity data.

shared_structs

sound_event

sound_events

Sound event registry entries matching vanilla Minecraft’s SoundEvents.

sound_types

Sound type definitions matching vanilla Minecraft’s SoundType.

stat

structure

timeline

trim_material

Armor trim material registry values.

trim_pattern

Armor trim pattern registry values.

vanilla_attributes

vanilla_banner_pattern_tags

vanilla_banner_patterns

vanilla_biome_tags

vanilla_biomes

vanilla_block_entity_types

vanilla_block_tags

vanilla_blocks

vanilla_cat_sound_variants

vanilla_cat_variants

vanilla_chat_types

vanilla_chicken_sound_variants

vanilla_chicken_variants

vanilla_configured_carvers

vanilla_configured_features

vanilla_cow_sound_variants

vanilla_cow_variants

vanilla_custom_stats

vanilla_damage_type_tags

vanilla_damage_types

vanilla_dialog_tags

vanilla_dialogs

vanilla_dimension_types

vanilla_enchantment_tags

vanilla_enchantments

vanilla_entities

vanilla_entity_data

vanilla_entity_type_tags

vanilla_fluid_tags

vanilla_fluids

vanilla_frog_variants

vanilla_game_events

vanilla_game_rules

vanilla_instrument_tags

vanilla_instruments

vanilla_item_tags

vanilla_items

vanilla_jukebox_songs

vanilla_loot_tables

vanilla_map_decoration_types

vanilla_menu_types

vanilla_mob_effects

vanilla_painting_variant_tags

vanilla_painting_variants

vanilla_particle_types

vanilla_pig_sound_variants

vanilla_pig_variants

vanilla_placed_features

vanilla_poi_type_tags

vanilla_poi_types

vanilla_position_source_types

vanilla_potion_tags

vanilla_potions

vanilla_recipes

vanilla_structure_processors

vanilla_structure_sets

vanilla_structure_tags

vanilla_structures

vanilla_template_pools

vanilla_timeline_tags

vanilla_timelines

vanilla_trim_materials

vanilla_trim_patterns

vanilla_villager_professions

vanilla_villager_types

vanilla_wolf_sound_variants

vanilla_wolf_variants

vanilla_world_clocks

vanilla_zombie_nautilus_variants

villager_profession

villager_type

wolf_sound_variant

wolf_variant

world_clock

zombie_nautilus_variant

Macros

impl_registry

Implements both RegistryExt and RegistryEntry for a standard registry.

impl_registry_entry

Implements RegistryEntry for an entry type via hash map lookup.

impl_registry_entry_eq

Implements registry-identity equality for an entry type by key.

impl_registry_ext

Implements RegistryExt for a registry type.

impl_standard_methods

Implements the default register and iter methods in the registries.

impl_tagged_registry

Implements TaggedRegistryExt for a registry with tag support.

offset

Macro to generate offset calculation from property values in all positions.