steel_registry

Module blocks

Modules

behavior

Block configuration and static properties.

block_state_ext

properties

shapes

Macros

offset

Re-export for easier access Macro to generate offset calculation from property values in all positions.

Structs

Block

BlockLightProperties

BlockRegistry

StateBooleanData

StateBooleanOverwrite

Functions

empty_shape 🔒

Default interaction shape function that returns an empty shape.

full_block_shape 🔒

Default shape function that returns a full block.

opaque_full_block_light_properties 🔒

Type Aliases

BlockRef

LightPropertiesFn

Function type for light-property lookups. Takes a state offset and returns extracted vanilla properties.

ShapeFn

Function type for shape lookups. Takes a state offset and returns the shape.