steel_registry::blocks::behaviour

Trait BlockBehaviour

pub trait BlockBehaviour: Send + Sync {
    // Required method
    fn get_state_for_placement(
        &self,
        context: &BlockPlaceContext<'_>,
    ) -> Option<BlockStateId>;

    // Provided methods
    fn update_shape(
        &self,
        state: BlockStateId,
        _world: &dyn RegistryWorld,
        _pos: BlockPos,
        _direction: Direction,
        _neighbor_pos: BlockPos,
        _neighbor_state: BlockStateId,
    ) -> BlockStateId { ... }
    fn on_place(
        &self,
        state: BlockStateId,
        world: &dyn RegistryWorld,
        pos: BlockPos,
        old_state: BlockStateId,
        moved_by_piston: bool,
    ) { ... }
    fn affect_neighbors_after_removal(
        &self,
        state: BlockStateId,
        world: &dyn RegistryWorld,
        pos: BlockPos,
        moved_by_piston: bool,
    ) { ... }
    fn use_item_on(
        &self,
        item_stack: &ItemStack,
        state: BlockStateId,
        world: &dyn RegistryWorld,
        pos: BlockPos,
        player: &dyn RegistryPlayer,
        hand: InteractionHand,
        hit_result: &BlockHitResult,
    ) -> InteractionResult { ... }
    fn use_without_item(
        &self,
        state: BlockStateId,
        world: &dyn RegistryWorld,
        pos: BlockPos,
        player: &dyn RegistryPlayer,
        hit_result: &BlockHitResult,
    ) -> InteractionResult { ... }
}

Required Methods

fn get_state_for_placement( &self, context: &BlockPlaceContext<'_>, ) -> Option<BlockStateId>

Returns the block state to use when placing this block.

Provided Methods

fn update_shape( &self, state: BlockStateId, _world: &dyn RegistryWorld, _pos: BlockPos, _direction: Direction, _neighbor_pos: BlockPos, _neighbor_state: BlockStateId, ) -> BlockStateId

Called when a neighboring block changes shape. Returns the new state for this block after considering the neighbor change.

fn on_place( &self, state: BlockStateId, world: &dyn RegistryWorld, pos: BlockPos, old_state: BlockStateId, moved_by_piston: bool, )

Called when this block is placed in the world.

Arguments

state - The new block state that was placed

world - The world the block was placed in

pos - The position where the block was placed

old_state - The previous block state at this position

moved_by_piston - Whether the block was moved by a piston

fn affect_neighbors_after_removal( &self, state: BlockStateId, world: &dyn RegistryWorld, pos: BlockPos, moved_by_piston: bool, )

Called after this block is removed from the world, to affect neighbors.

This is used for things like rails notifying neighbors when removed.

Arguments

state - The block state that was removed

world - The world the block was removed from

pos - The position where the block was removed

moved_by_piston - Whether the block was moved by a piston

fn use_item_on( &self, item_stack: &ItemStack, state: BlockStateId, world: &dyn RegistryWorld, pos: BlockPos, player: &dyn RegistryPlayer, hand: InteractionHand, hit_result: &BlockHitResult, ) -> InteractionResult

Called when a player uses an item on this block.

Returns TryEmptyHandInteraction by default to fall through to item use. Override this to handle block-specific interactions (e.g., opening chests, using buttons, etc.).

fn use_without_item( &self, state: BlockStateId, world: &dyn RegistryWorld, pos: BlockPos, player: &dyn RegistryPlayer, hit_result: &BlockHitResult, ) -> InteractionResult

Called when a player uses this block without an item (or as a fallback when use_item_on returns TryEmptyHandInteraction).

Returns Pass by default. Override this for blocks that have interactions without needing an item (e.g., buttons, levers, repeaters).