steel_registry::data_components

Module registry

Data component registry and storage types.

This module provides:

DataComponentRegistry - Registry of all component types with their serialization functions

DataComponentMap - Storage for component values on items/entities

DataComponentPatch - Diff representation for network/storage

DataComponentType - Type-safe handle for accessing components

Structs

ComponentEntry

Metadata for a registered component type.

DataComponentMap

Storage for component values.

DataComponentPatch

A patch representing modifications to a DataComponentMap.

DataComponentRegistry

Registry of all data component types.

DataComponentType

A typed handle for a data component.

Enums

ComponentPatchEntry

Entry in a component patch.

Functions

component_try_into

Attempts to extract a typed component from ComponentData.

Type Aliases

ComponentEntryRef

NbtReader

Reader function for deserializing a component from NBT format.

NbtWriter

Writer function for serializing a component to NBT format.

NetworkReader

Reader function for deserializing a component from network format.

NetworkWriter

Writer function for serializing a component to network format.