steel_registry::data_components::registry

Module codecs

Structs

ComponentCodecs 🔒

ComponentEntry

Metadata for a registered component type.

NetworkCodecs 🔒

PersistentCodecs 🔒

Traits

ValidatePersistentComponent 🔒

Additional source-value validation required before persistent encoding.

Functions

hash_component 🔒

read_typed_nbt 🔒

read_typed_network 🔒

validate_component 🔒

write_typed_nbt 🔒

write_typed_network 🔒

Type Aliases

ComponentEntryRef

ComponentHash 🔒

Function for hashing a component through its persistent codec shape.

ComponentValidator 🔒

NbtReader

Reader function for deserializing a component from NBT format.

NbtWriter

Writer function for serializing a component to NBT format.

NetworkReader

NetworkWriter

Writer function for serializing a component to network format.

PersistentCodecFns 🔒