steel_registry::data_components

Struct DataComponentRegistry

pub struct DataComponentRegistry {
    components_by_key: FxHashMap<Identifier, usize>,
    components_by_id: Vec<Identifier>,
    readers_by_id: Vec<fn(&mut Cursor<&[u8]>) -> Result<Box<dyn ComponentValue>>>,
    nbt_readers_by_id: Vec<fn(NbtTag<'_, '_>) -> Option<Box<dyn ComponentValue>>>,
    allows_registering: bool,
}

Fields
components_by_key: FxHashMap<Identifier, usize>

Map from component key to network ID

components_by_id: Vec<Identifier>

Ordered list of component keys (index = network ID)

readers_by_id: Vec<fn(&mut Cursor<&[u8]>) -> Result<Box<dyn ComponentValue>>>

Ordered list of network reader functions (index = network ID)

nbt_readers_by_id: Vec<fn(NbtTag<'_, '_>) -> Option<Box<dyn ComponentValue>>>

Ordered list of NBT reader functions (index = network ID)

allows_registering: bool
Implementations

impl DataComponentRegistry

pub fn new() -> Self

pub fn register<T: 'static + ComponentValue + ReadFrom + FromNbtTag>( &mut self, component: DataComponentType<T>, )

pub fn get_id<T: 'static>( &self, component: DataComponentType<T>, ) -> Option<usize>

pub fn get_id_by_key(&self, key: &Identifier) -> Option<usize>

pub fn get_key_by_id(&self, id: usize) -> Option<&Identifier>

Gets the component key by its network ID.

pub fn get_reader_by_id( &self, id: usize, ) -> Option<fn(&mut Cursor<&[u8]>) -> Result<Box<dyn ComponentValue>>>

Gets the network reader function for a component by its network ID.

pub fn get_nbt_reader_by_id( &self, id: usize, ) -> Option<fn(NbtTag<'_, '_>) -> Option<Box<dyn ComponentValue>>>

Gets the NBT reader function for a component by its network ID.

pub fn get_nbt_reader_by_key( &self, key: &Identifier, ) -> Option<fn(NbtTag<'_, '_>) -> Option<Box<dyn ComponentValue>>>

Gets the NBT reader function for a component by its key.

pub fn len(&self) -> usize

Returns the number of registered components.

pub fn is_empty(&self) -> bool

Returns true if no components are registered.