steel_registry::data_components

Module vanilla_components

Vanilla data component definitions and registration.

This module defines all vanilla Minecraft data components and provides the registration function to add them to the registry.

Re-exports

pub use crate::attribute::AttributeModifierOperation;

pub use crate::equipment::EquipmentSlot;

pub use crate::equipment::EquipmentSlotGroup;

pub use super::components::AttackRange;

pub use super::components::DamageTypeComponent;

pub use super::components::Equippable;

pub use super::components::EquippableAllowedEntities;

pub use super::components::ItemAttributeModifierDisplay;

pub use super::components::ItemAttributeModifierEntry;

pub use super::components::ItemAttributeModifiers;

pub use super::components::ItemEnchantments;

pub use super::components::PiercingWeapon;

pub use super::components::Tool;

pub use super::components::ToolRule;

pub use super::components::Weapon;

Macros

register_stub 🔒

Helper to create stub reader/writer functions for unimplemented components. These components use the Todo variant as a placeholder.

Structs

DataComponentType

A typed handle for a data component.

Constants

ADDITIONAL_TRADE_COST

ATTACK_RANGE

ATTRIBUTE_MODIFIERS

AXOLOTL_VARIANT

BANNER_PATTERNS

BASE_COLOR

BEES

BLOCKS_ATTACKS

BLOCK_ENTITY_DATA

BLOCK_STATE

BREAK_SOUND

BUCKET_ENTITY_DATA

BUNDLE_CONTENTS

CAN_BREAK

CAN_PLACE_ON

CAT_COLLAR

CAT_SOUND_VARIANT

CAT_VARIANT

CHARGED_PROJECTILES

CHICKEN_SOUND_VARIANT

CHICKEN_VARIANT

CONSUMABLE

CONTAINER

CONTAINER_LOOT

COW_SOUND_VARIANT

COW_VARIANT

CREATIVE_SLOT_LOCK

CUSTOM_DATA

CUSTOM_MODEL_DATA

CUSTOM_NAME

DAMAGE

DAMAGE_RESISTANT

DAMAGE_TYPE

DEATH_PROTECTION

DEBUG_STICK_STATE

DYE

DYED_COLOR

ENCHANTABLE

ENCHANTMENTS

ENCHANTMENT_GLINT_OVERRIDE

ENTITY_DATA

EQUIPPABLE

FIREWORKS

FIREWORK_EXPLOSION

FOOD

FOX_VARIANT

FROG_VARIANT

GLIDER

HORSE_VARIANT

INSTRUMENT

INTANGIBLE_PROJECTILE

ITEM_MODEL

ITEM_NAME

JUKEBOX_PLAYABLE

KINETIC_WEAPON

LLAMA_VARIANT

LOCK

LODESTONE_TRACKER

LORE

MAP_COLOR

MAP_DECORATIONS

MAP_ID

MAP_POST_PROCESSING

MAX_DAMAGE

MAX_STACK_SIZE

MINIMUM_ATTACK_CHARGE

MOOSHROOM_VARIANT

NOTE_BLOCK_SOUND

OMINOUS_BOTTLE_AMPLIFIER

PAINTING_VARIANT

PARROT_VARIANT

PIERCING_WEAPON

PIG_SOUND_VARIANT

PIG_VARIANT

POTION_CONTENTS

POTION_DURATION_SCALE

POT_DECORATIONS

PROFILE

PROVIDES_BANNER_PATTERNS

PROVIDES_TRIM_MATERIAL

RABBIT_VARIANT

RARITY

RECIPES

REPAIRABLE

REPAIR_COST

SALMON_SIZE

SHEEP_COLOR

SHULKER_COLOR

STORED_ENCHANTMENTS

SULFUR_CUBE_CONTENT

SUSPICIOUS_STEW_EFFECTS

SWING_ANIMATION

TOOL

TOOLTIP_DISPLAY

TOOLTIP_STYLE

TRIM

TROPICAL_FISH_BASE_COLOR

TROPICAL_FISH_PATTERN

TROPICAL_FISH_PATTERN_COLOR

UNBREAKABLE

USE_COOLDOWN

USE_EFFECTS

USE_REMAINDER

VILLAGER_VARIANT

WEAPON

WOLF_COLLAR

WOLF_SOUND_VARIANT

WOLF_VARIANT

WRITABLE_BOOK_CONTENT

WRITTEN_BOOK_CONTENT

ZOMBIE_NAUTILUS_VARIANT

Functions

register_vanilla_data_components

Registers all vanilla data components.

varint_reader 🔒

Network reader for VarInt-encoded i32 components.

varint_writer 🔒

Network writer for VarInt-encoded i32 components.