steel_registry::enchantment

Struct Enchantment

pub struct Enchantment {
    pub key: Identifier,
    pub max_level: u32,
    pub min_cost: EnchantmentCost,
    pub max_cost: EnchantmentCost,
    pub anvil_cost: i32,
    pub weight: u32,
    pub slots: &'static [EquipmentSlotGroup],
    pub supported_items: &'static str,
    pub primary_items: Option<&'static str>,
    pub exclusive_set: Option<&'static str>,
    pub effects_nbt: fn() -> NbtCompound,
    pub effects: EnchantmentEffects,
}

Fields
key: Identifiermax_level: u32min_cost: EnchantmentCostmax_cost: EnchantmentCostanvil_cost: i32weight: u32slots: &'static [EquipmentSlotGroup]supported_items: &'static strprimary_items: Option<&'static str>exclusive_set: Option<&'static str>effects_nbt: fn() -> NbtCompoundeffects: EnchantmentEffects
Implementations

impl Enchantment

pub fn matching_slot(&self, slot: EquipmentSlot) -> bool

Vanilla Enchantment::matchingSlot.

pub fn can_enchant(&self, item: ItemRef) -> bool

Checks if this enchantment can be applied to the given item via supported_items tag.

pub fn are_compatible(a: EnchantmentRef, b: EnchantmentRef) -> bool

Checks if two enchantments are compatible (neither’s exclusive_set contains the other).

pub fn is_compatible_with_existing( enchantment: EnchantmentRef, item: &ItemStack, ) -> bool

Checks if this enchantment is compatible with all existing enchantments on an item.