steel_registry::vanilla_entity_data

Trait VanillaEntityData

pub trait VanillaEntityData {
    // Required methods
    fn base(&self) -> &BaseEntityData;
    fn base_mut(&mut self) -> &mut BaseEntityData;
    fn pack_dirty(&mut self) -> Option<Vec<DataValue>>;
    fn pack_all(&self) -> Vec<DataValue>;
    fn is_dirty(&self) -> bool;
}

Common access to the vanilla synchronized entity data root layer.

Required Methods

fn base(&self) -> &BaseEntityData

Returns the shared vanilla base entity-data layer.

fn base_mut(&mut self) -> &mut BaseEntityData

Returns the mutable shared vanilla base entity-data layer.

fn pack_dirty(&mut self) -> Option<Vec<DataValue>>

Packs dirty values for network sync, clearing dirty flags.

fn pack_all(&self) -> Vec<DataValue>

Packs all non-default values for initial entity spawn.

fn is_dirty(&self) -> bool

Returns true if any field has been modified.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".