steel_registry

Module vanilla_template_pools

Constants

POOL_COUNT

Number of template pools.

TEMPLATE_COUNT

Number of structure templates.

Functions

vanilla_template_nbt_bytes

Returns the compressed NBT bytes for a vanilla structure template.

vanilla_template_pools

Returns all vanilla template pools parsed from the datapack.

vanilla_templates

Returns all vanilla structure templates with their jigsaw data.