Crate steel_registry

Modules

attribute

banner_pattern

biome

block_entity_type

blocks

carver

Configured carver registry.

cat_sound_variant

cat_variant

chat_type

chicken_sound_variant

chicken_variant

cow_sound_variant

cow_variant

damage_type

data_components

Data components system for items and entities.

dialog

dimension_type

enchantment

enchantment_effect

entity_data

Entity data synchronization system.

entity_type

equipment

Shared equipment slot definitions.

feature

Configured and placed feature registries.

fluid

Fluid registry for Minecraft fluids.

frog_variant

game_events

game_rules

instrument

item_stack

Item stack implementation.

items

jukebox_song

level_events

Level event constants matching vanilla Minecraft’s LevelEvent.java.

loot_table

macros 🔒

menu_type

mob_effect

packets

painting_variant

particle_type

pig_sound_variant

pig_variant

poi

Point of Interest (POI) type registry.

recipe

Recipe system for crafting and other recipe types.

shared_structs

sound_event

sound_events

Sound event registry entries matching vanilla Minecraft’s SoundEvents.

sound_types

Sound type definitions matching vanilla Minecraft’s SoundType.

structure

structure_processor

Structure processor-list registry data.

structure_set

Structure set data types for generated registry data.

template_pool

Template pool and structure template data for jigsaw assembly.

test_support

timeline

trim_material

trim_pattern

vanilla_attributes

vanilla_banner_pattern_tags

vanilla_banner_patterns

vanilla_biome_tags

vanilla_biomes

vanilla_block_entity_types

vanilla_block_tags

vanilla_blocks

vanilla_cat_sound_variants

vanilla_cat_variants

vanilla_chat_types

vanilla_chicken_sound_variants

vanilla_chicken_variants

vanilla_configured_carvers

vanilla_configured_features

vanilla_cow_sound_variants

vanilla_cow_variants

vanilla_damage_type_tags

vanilla_damage_types

vanilla_dialog_tags

vanilla_dialogs

vanilla_dimension_types

vanilla_enchantment_tags

vanilla_enchantments

vanilla_entities

vanilla_entity_data

vanilla_entity_type_tags

vanilla_fluid_tags

vanilla_fluids

vanilla_frog_variants

vanilla_game_events

vanilla_game_rules

vanilla_instrument_tags

vanilla_instruments

vanilla_item_tags

vanilla_items

vanilla_jukebox_songs

vanilla_loot_tables

vanilla_menu_types

vanilla_mob_effects

vanilla_painting_variant_tags

vanilla_painting_variants

vanilla_particle_types

vanilla_pig_sound_variants

vanilla_pig_variants

vanilla_placed_features

vanilla_poi_type_tags

vanilla_poi_types

vanilla_recipes

vanilla_structure_processors

vanilla_structure_sets

vanilla_structure_tags

vanilla_structures

vanilla_template_pools

vanilla_timeline_tags

vanilla_timelines

vanilla_trim_materials

vanilla_trim_patterns

vanilla_villager_professions

vanilla_villager_types

vanilla_wolf_sound_variants

vanilla_wolf_variants

vanilla_world_clocks

vanilla_zombie_nautilus_variants

villager_profession

villager_type

wolf_sound_variant

wolf_variant

world_clock

zombie_nautilus_variant

Macros

impl_registry

Implements both RegistryExt and RegistryEntry for a standard registry.

impl_registry_entry

Implements RegistryEntry for an entry type via hash map lookup.

impl_registry_entry_eq

Implements registry-identity equality for an entry type by key.

impl_registry_ext

Implements RegistryExt for a registry type.

impl_standard_methods

Implements the default register, replace, and iter methods in the registries

impl_tagged_registry

Implements TaggedRegistryExt for a registry with tag support.

offset

Macro to generate offset calculation from property values in all positions.

Structs

Registry

RegistryLock

Constants

BANNER_PATTERN_REGISTRY

BIOMES_REGISTRY

BLOCKS_REGISTRY

BLOCK_ENTITY_TYPE_REGISTRY

CAT_SOUND_VARIANT_REGISTRY

CAT_VARIANT_REGISTRY

CHAT_TYPE_REGISTRY

CHICKEN_SOUND_VARIANT_REGISTRY

CHICKEN_VARIANT_REGISTRY

CONFIGURED_CARVER_REGISTRY

CONFIGURED_FEATURE_REGISTRY

COW_SOUND_VARIANT_REGISTRY

COW_VARIANT_REGISTRY

DAMAGE_TYPE_REGISTRY

DIALOG_REGISTRY

DIMENSION_TYPE_REGISTRY

ENCHANTMENT_REGISTRY

ENTITY_TYPE_REGISTRY

FLUID_REGISTRY

FROG_VARIANT_REGISTRY

INSTRUMENT_REGISTRY

ITEMS_REGISTRY

JUKEBOX_SONG_REGISTRY

LOOT_TABLE_REGISTRY

MENU_TYPE_REGISTRY

MOB_EFFECT_REGISTRY

PAINTING_VARIANT_REGISTRY

PARTICLE_TYPE_REGISTRY

PIG_SOUND_VARIANT_REGISTRY

PIG_VARIANT_REGISTRY

PLACED_FEATURE_REGISTRY

POI_TYPE_REGISTRY

SOUND_EVENT_REGISTRY

STRUCTURE_PROCESSOR_LIST_REGISTRY

STRUCTURE_REGISTRY

TIMELINE_REGISTRY

TRIM_MATERIAL_REGISTRY

TRIM_PATTERN_REGISTRY

VILLAGER_PROFESSION_REGISTRY

VILLAGER_TYPE_REGISTRY

WOLF_SOUND_VARIANT_REGISTRY

WOLF_VARIANT_REGISTRY

WORLD_CLOCK_REGISTRY

ZOMBIE_NAUTILUS_VARIANT_REGISTRY

Statics

REGISTRY

Traits

RegistryEntry

Trait for types stored in a registry, allowing self-lookup of their numeric ID.

RegistryExt

Generic trait for registries with a typed entry.

TaggedRegistryExt

Trait for registries that support tagging entries.