steel_registry::blocks::block_state_ext

Trait BlockStateExt

pub trait BlockStateExt {
    // Required methods
    fn get_block(&self) -> BlockRef;
    fn is_air(&self) -> bool;
    fn has_block_entity(&self) -> bool;
    fn get_value<T, P: Property<T>>(&self, property: &P) -> T;
    fn try_get_value<T, P: Property<T>>(&self, property: &P) -> Option<T>;
    fn set_value<T, P: Property<T>>(
        &self,
        property: &P,
        value: T,
    ) -> BlockStateId;
    fn get_property_str(&self, name: &str) -> Option<String>;
    fn get_collision_shape(&self) -> &'static [AABB];
    fn get_outline_shape(&self) -> &'static [AABB];
    fn is_face_sturdy(&self, direction: Direction) -> bool;
    fn is_face_sturdy_for(
        &self,
        direction: Direction,
        support_type: SupportType,
    ) -> bool;
    fn is_solid(&self) -> bool;
}

Required Methods

fn get_block(&self) -> BlockRef

fn is_air(&self) -> bool

fn has_block_entity(&self) -> bool

fn get_value<T, P: Property<T>>(&self, property: &P) -> T

fn try_get_value<T, P: Property<T>>(&self, property: &P) -> Option<T>

Gets the value of a property, returning None if the block doesn’t have this property.

fn set_value<T, P: Property<T>>(&self, property: &P, value: T) -> BlockStateId

fn get_property_str(&self, name: &str) -> Option<String>

fn get_collision_shape(&self) -> &'static [AABB]

fn get_outline_shape(&self) -> &'static [AABB]

fn is_face_sturdy(&self, direction: Direction) -> bool

Checks if this block face is sturdy enough to support other blocks. Uses SupportType::Full by default.

fn is_face_sturdy_for( &self, direction: Direction, support_type: SupportType, ) -> bool

Checks if this block face is sturdy for the given support type.

fn is_solid(&self) -> bool

Checks if this block state is solid (has a full cube collision shape).

This matches vanilla’s BlockState.isSolid() which is used by standing signs to check if they can be placed on a block.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.