steel_registry::blocks

Struct BlockRegistry

pub struct BlockRegistry {
    blocks_by_id: Vec<BlockRef>,
    blocks_by_key: FxHashMap<Identifier, usize>,
    tags: FxHashMap<Identifier, Vec<BlockRef>>,
    allows_registering: bool,
    pub state_to_block_lookup: Vec<BlockRef>,
    pub state_to_block_id: Vec<usize>,
    pub block_to_base_state: Vec<u16>,
    pub next_state_id: u16,
}

Fields
blocks_by_id: Vec<BlockRef>blocks_by_key: FxHashMap<Identifier, usize>tags: FxHashMap<Identifier, Vec<BlockRef>>allows_registering: boolstate_to_block_lookup: Vec<BlockRef>state_to_block_id: Vec<usize>

Maps state IDs to block IDs (parallel to state_to_block_lookup for O(1) lookup)

block_to_base_state: Vec<u16>

Maps block IDs to their base state ID

next_state_id: u16

The next state ID to be allocated

Implementations

impl BlockRegistry

pub fn new() -> Self

pub fn register(&mut self, block: BlockRef) -> usize

pub fn get_base_state_id(&self, block: BlockRef) -> BlockStateId

pub fn get_default_state_id(&self, block: BlockRef) -> BlockStateId

Gets the default state ID for a block (base state + default offset)

pub fn by_id(&self, id: usize) -> Option<BlockRef>

pub fn get_id(&self, block: BlockRef) -> &usize

pub fn by_state_id(&self, state_id: BlockStateId) -> Option<BlockRef>

pub fn by_key(&self, key: &Identifier) -> Option<BlockRef>

pub fn get_properties(&self, id: BlockStateId) -> Vec<(&str, &str)>

pub fn state_id_from_properties( &self, key: &Identifier, properties: &[(&str, &str)], ) -> Option<BlockStateId>

Gets the state ID for a block with the given properties.

Returns None if the block key is unknown or if any property name/value is invalid.

Properties can be provided in any order. Missing properties will use the block’s default values (typically index 0 for each property).

pub fn get_property<T, P: Property<T>>( &self, id: BlockStateId, property: &P, ) -> T

pub fn try_get_property<T, P: Property<T>>( &self, id: BlockStateId, property: &P, ) -> Option<T>

Gets the value of a property, returning None if the block doesn’t have this property.

pub fn set_property<T, P: Property<T>>( &self, id: BlockStateId, property: &P, value: T, ) -> BlockStateId

pub fn iter(&self) -> impl Iterator<Item = (usize, BlockRef)> + '_

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn register_tag(&mut self, tag: Identifier, block_keys: &[&'static str])

Registers a tag with a list of block keys. Block keys that don’t exist in the registry are silently skipped.

pub fn is_in_tag(&self, block: BlockRef, tag: &Identifier) -> bool

Checks if a block is in a given tag.

pub fn get_tag(&self, tag: &Identifier) -> Option<&[BlockRef]>

Gets all blocks in a tag.

pub fn iter_tag(&self, tag: &Identifier) -> impl Iterator<Item = BlockRef> + '_

Iterates over all blocks in a tag.

pub fn tag_keys(&self) -> impl Iterator<Item = &Identifier> + '_

Gets all tag keys.

impl BlockRegistry

pub fn get_collision_shape(&self, state_id: BlockStateId) -> &'static [AABB]

Gets the collision shape for a block state.

Returns a slice of AABBs that make up the collision shape. For simple blocks this is typically a single full-block AABB. For complex blocks like fences, this may be multiple AABBs.

pub fn get_outline_shape(&self, state_id: BlockStateId) -> &'static [AABB]

Gets the outline shape for a block state.

This is the shape shown when the player targets the block. Often the same as collision shape, but can differ (e.g., fences).

pub fn get_shapes(&self, state_id: BlockStateId) -> BlockShapes

Gets both collision and outline shapes for a block state.