steel_registry::data_components

Module vanilla_components

Vanilla data component definitions and registration.

This module defines all vanilla Minecraft data components and provides the registration function to add them to the registry.

Re-exports

pub use crate::attribute::AttributeModifierOperation;

pub use crate::equipment::EquipmentSlot;

pub use crate::equipment::EquipmentSlotGroup;

pub use super::components::ArmorTrim;

pub use super::components::AttackRange;

pub use super::components::BannerPatternLayer;

pub use super::components::BannerPatternLayers;

pub use super::components::BeehiveOccupant;

pub use super::components::Bees;

pub use super::components::BlockEntityData;

pub use super::components::BlockItemStateProperties;

pub use super::components::BlocksAttacks;

pub use super::components::BundleContents;

pub use super::components::ChargedProjectiles;

pub use super::components::Consumable;

pub use super::components::CustomData;

pub use super::components::CustomModelData;

pub use super::components::DamageReduction;

pub use super::components::DamageResistant;

pub use super::components::DamageTypeComponent;

pub use super::components::DeathProtection;

pub use super::components::DebugStickProperty;

pub use super::components::DebugStickState;

pub use super::components::DyedItemColor;

pub use super::components::Enchantable;

pub use super::components::EntityData;

pub use super::components::Equippable;

pub use super::components::EquippableAllowedEntities;

pub use super::components::Filterable;

pub use super::components::FireworkExplosion;

pub use super::components::FireworkExplosionShape;

pub use super::components::Fireworks;

pub use super::components::FoodProperties;

pub use super::components::GlobalPos;

pub use super::components::InstrumentComponent;

pub use super::components::InvalidEnchantableValue;

pub use super::components::ItemAttributeModifierDisplay;

pub use super::components::ItemAttributeModifierEntry;

pub use super::components::ItemAttributeModifiers;

pub use super::components::ItemContainerContents;

pub use super::components::ItemDamageFunction;

pub use super::components::ItemEnchantments;

pub use super::components::ItemLore;

pub use super::components::ItemLoreTooLong;

pub use super::components::ItemUseAnimation;

pub use super::components::JukeboxPlayable;

pub use super::components::KineticWeapon;

pub use super::components::KineticWeaponCondition;

pub use super::components::LodestoneTracker;

pub use super::components::MapDecorationEntry;

pub use super::components::MapDecorations;

pub use super::components::MapId;

pub use super::components::MapItemColor;

pub use super::components::MapPostProcessing;

pub use super::components::OminousBottleAmplifier;

pub use super::components::PaintingVariantComponent;

pub use super::components::PiercingWeapon;

pub use super::components::PotDecorations;

pub use super::components::PotionContents;

pub use super::components::ProvidesBannerPatterns;

pub use super::components::ProvidesTrimMaterial;

pub use super::components::Rarity;

pub use super::components::Recipes;

pub use super::components::Repairable;

pub use super::components::SeededContainerLoot;

pub use super::components::SulfurCubeContent;

pub use super::components::SuspiciousStewEffect;

pub use super::components::SuspiciousStewEffects;

pub use super::components::SwingAnimation;

pub use super::components::SwingAnimationType;

pub use super::components::Tool;

pub use super::components::ToolRule;

pub use super::components::ToolRuleBlocks;

pub use super::components::TooltipDisplay;

pub use super::components::UseCooldown;

pub use super::components::UseEffects;

pub use super::components::UseRemainder;

pub use super::components::Weapon;

pub use super::components::WritableBookContent;

pub use super::components::WrittenBookContent;

pub use crate::ItemStackTemplate;

pub use crate::cat_sound_variant::CatSoundVariant;

pub use crate::cat_variant::CatVariant;

pub use crate::chicken_sound_variant::ChickenSoundVariant;

pub use crate::chicken_variant::ChickenVariant;

pub use crate::cow_sound_variant::CowSoundVariant;

pub use crate::cow_variant::CowVariant;

pub use crate::frog_variant::FrogVariant;

pub use crate::item_predicate::AdventureModePredicate;

pub use crate::item_predicate::BlockPredicate;

pub use crate::item_predicate::ItemPredicate;

pub use crate::item_predicate::LockCode;

pub use crate::pig_sound_variant::PigSoundVariant;

pub use crate::pig_variant::PigVariant;

pub use crate::resolvable_profile::PartialProfile;

pub use crate::resolvable_profile::PlayerModelType;

pub use crate::resolvable_profile::PlayerSkinPatch;

pub use crate::resolvable_profile::ProfileProperty;

pub use crate::resolvable_profile::ResolvableProfile;

pub use crate::resolvable_profile::ResolvableProfileContents;

pub use crate::resolvable_profile::StoredGameProfile;

pub use crate::sound_event::SoundEventHolder;

pub use crate::villager_type::VillagerType;

pub use crate::wolf_sound_variant::WolfSoundVariant;

pub use crate::wolf_variant::WolfVariant;

pub use crate::zombie_nautilus_variant::ZombieNautilusVariant;

pub use crate::AxolotlVariant;

pub use crate::DyeColor;

pub use crate::FoxVariant;

pub use crate::HorseVariant;

pub use crate::LlamaVariant;

pub use crate::MooshroomVariant;

pub use crate::ParrotVariant;

pub use crate::RabbitVariant;

pub use crate::RegistryReference;

pub use crate::SalmonVariant;

pub use crate::TropicalFishPattern;

Macros

register_ranged_i32 🔒

register_stream_unit 🔒

Structs

DataComponentType

A typed handle for a data component.

Constants

ADDITIONAL_TRADE_COST

ATTACK_RANGE

ATTRIBUTE_MODIFIERS

AXOLOTL_VARIANT

BANNER_PATTERNS

BASE_COLOR

BEES

BLOCKS_ATTACKS

BLOCK_ENTITY_DATA

BLOCK_STATE

BREAK_SOUND

BUCKET_ENTITY_DATA

BUNDLE_CONTENTS

CAN_BREAK

CAN_PLACE_ON

CAT_COLLAR

CAT_SOUND_VARIANT

CAT_VARIANT

CHARGED_PROJECTILES

CHICKEN_SOUND_VARIANT

CHICKEN_VARIANT

CONSUMABLE

CONTAINER

CONTAINER_LOOT

COW_SOUND_VARIANT

COW_VARIANT

CREATIVE_SLOT_LOCK

CUSTOM_DATA

CUSTOM_MODEL_DATA

CUSTOM_NAME

DAMAGE

DAMAGE_RESISTANT

DAMAGE_TYPE

DEATH_PROTECTION

DEBUG_STICK_STATE

DYE

DYED_COLOR

ENCHANTABLE

ENCHANTMENTS

ENCHANTMENT_GLINT_OVERRIDE

ENTITY_DATA

EQUIPPABLE

FIREWORKS

FIREWORK_EXPLOSION

FOOD

FOX_VARIANT

FROG_VARIANT

GLIDER

HORSE_VARIANT

INSTRUMENT

INTANGIBLE_PROJECTILE

ITEM_MODEL

ITEM_NAME

JUKEBOX_PLAYABLE

KINETIC_WEAPON

LLAMA_VARIANT

LOCK

LODESTONE_TRACKER

LORE

MAP_COLOR

MAP_DECORATIONS

MAP_ID

MAP_POST_PROCESSING

MAX_DAMAGE

MAX_STACK_SIZE

MINIMUM_ATTACK_CHARGE

MOOSHROOM_VARIANT

NOTE_BLOCK_SOUND

OMINOUS_BOTTLE_AMPLIFIER

PAINTING_VARIANT

PARROT_VARIANT

PIERCING_WEAPON

PIG_SOUND_VARIANT

PIG_VARIANT

POTION_CONTENTS

POTION_DURATION_SCALE

POT_DECORATIONS

PROFILE

PROVIDES_BANNER_PATTERNS

PROVIDES_TRIM_MATERIAL

RABBIT_VARIANT

RARITY

RECIPES

REPAIRABLE

REPAIR_COST

SALMON_SIZE

SHEEP_COLOR

SHULKER_COLOR

STORED_ENCHANTMENTS

SULFUR_CUBE_CONTENT

SUSPICIOUS_STEW_EFFECTS

SWING_ANIMATION

TOOL

TOOLTIP_DISPLAY

TOOLTIP_STYLE

TRIM

TROPICAL_FISH_BASE_COLOR

TROPICAL_FISH_PATTERN

TROPICAL_FISH_PATTERN_COLOR

UNBREAKABLE

USE_COOLDOWN

USE_EFFECTS

USE_REMAINDER

VILLAGER_VARIANT

WEAPON

WOLF_COLLAR

WOLF_SOUND_VARIANT

WOLF_VARIANT

WRITABLE_BOOK_CONTENT

WRITTEN_BOOK_CONTENT

ZOMBIE_NAUTILUS_VARIANT

Functions

bool_nbt_reader 🔒

bool_nbt_writer 🔒

bool_reader 🔒

bool_writer 🔒

codec_unit_network_reader 🔒

codec_unit_network_writer 🔒

custom_data_codec_reader 🔒

custom_data_writer 🔒

fireworks_nbt_reader 🔒

fireworks_nbt_writer 🔒

fireworks_network_reader 🔒

fireworks_network_writer 🔒

float_reader 🔒

float_writer 🔒

jukebox_playable_nbt_reader 🔒

jukebox_playable_nbt_writer 🔒

jukebox_playable_network_reader 🔒

jukebox_playable_network_writer 🔒

minimum_attack_charge_nbt_reader 🔒

minimum_attack_charge_nbt_writer 🔒

painting_variant_nbt_reader 🔒

painting_variant_nbt_writer 🔒

painting_variant_network_reader 🔒

painting_variant_network_writer 🔒

potion_duration_scale_nbt_reader 🔒

potion_duration_scale_nbt_writer 🔒

ranged_i32_nbt_reader 🔒

ranged_i32_nbt_writer 🔒

register_vanilla_data_components

Registers all vanilla data components.

text_component_nbt_reader 🔒

text_component_nbt_writer 🔒

text_component_network_reader 🔒

text_component_network_writer 🔒

unit_nbt_reader 🔒

unit_nbt_writer 🔒

unit_reader 🔒

unit_writer 🔒

varint_reader 🔒

Network reader for VarInt-encoded i32 components.

varint_writer 🔒

Network writer for VarInt-encoded i32 components.