steel_registry::item_stack

Struct ItemStack

pub struct ItemStack {
    pub item: ItemRef,
    pub count: i32,
    pub patch: DataComponentPatch,
}

A stack of items with a count and component modifications.

Fields
item: ItemRef

The item type. AIR represents an empty stack.

count: i32

The number of items in this stack.

patch: DataComponentPatch

Modifications to the prototype components.

Implementations

impl ItemStack

pub fn empty() -> Self

Creates an empty item stack (using AIR).

pub fn new(item: ItemRef) -> Self

Creates a new item stack with count 1.

pub fn with_count(item: ItemRef, count: i32) -> Self

Creates a new item stack with the specified count.

pub fn prototype(&self) -> &'static DataComponentMap

pub fn is_empty(&self) -> bool

pub fn item(&self) -> ItemRef

pub fn count(&self) -> i32

pub fn set_count(&mut self, count: i32)

pub fn grow(&mut self, amount: i32)

Increases the count by the given amount.

pub fn shrink(&mut self, amount: i32)

Decreases the count by the given amount.

pub fn is_stackable(&self) -> bool

Returns true if this item can stack (max stack size > 1 and not damaged). Damaged items cannot stack.

pub fn is_damageable_item(&self) -> bool

Returns true if this item can take damage.

pub fn is_damaged(&self) -> bool

Returns true if this item has taken damage.

pub fn get_damage_value(&self) -> i32

Gets the current damage value of this item.

pub fn get_max_damage(&self) -> i32

Gets the maximum damage this item can take before breaking.

pub fn has_component(&self, key: &Identifier) -> bool

Returns true if this item has the specified component.

pub fn is_same_item(a: &Self, b: &Self) -> bool

pub fn is_same_item_same_components(a: &Self, b: &Self) -> bool

Checks if two stacks have the same item and components.

pub fn matches(a: &Self, b: &Self) -> bool

pub fn is(&self, item: ItemRef) -> bool

pub fn max_stack_size(&self) -> i32

pub fn get_equippable(&self) -> Option<&Equippable>

Returns the equippable component if this item has one.

pub fn get_equippable_slot(&self) -> Option<EquippableSlot>

Returns the equipment slot this item can be equipped to, if any.

pub fn is_equippable_in_slot(&self, slot: EquippableSlot) -> bool

Returns true if this item can be equipped in the given slot.

pub fn get_effective_value_raw( &self, key: &Identifier, ) -> Option<&dyn ComponentValue>

pub fn components_equal(&self, other: &Self) -> bool