steel_registry::item_stack

Struct ItemStack

pub struct ItemStack {
    item: ItemRef,
    count: i32,
    patch: DataComponentPatch,
}

A stack of items with a count and component modifications.

Fields
item: ItemRef

The item type. AIR represents an empty stack.

count: i32

The number of items in this stack.

patch: DataComponentPatch

Modifications to the prototype components.

Implementations

impl ItemStack

pub fn empty() -> Self

Creates an empty item stack (using AIR).

pub fn new(item: ItemRef) -> Self

Creates a new item stack with count 1.

pub fn with_count(item: ItemRef, count: i32) -> Self

Creates a new item stack with the specified count.

pub fn prototype(&self) -> &'static DataComponentMap

pub fn is_empty(&self) -> bool

pub fn item(&self) -> ItemRef

pub fn count(&self) -> i32

pub fn set_count(&mut self, count: i32)

pub fn is_same_item(a: &Self, b: &Self) -> bool

pub fn is_same_item_same_components(a: &Self, b: &Self) -> bool

Checks if two stacks have the same item and components.

pub fn matches(a: &Self, b: &Self) -> bool

pub fn is(&self, item: ItemRef) -> bool